Make native_server games seedable
Uses the same algorithm and the JS server, so seeds are (for now) compatible

+++ src/server/json_api.rs
@@ -63,7 +63,7 @@ pub mod resp {
     #[serde(rename_all = "camelCase")]
     pub struct ServerResponse {
         pub id: String,
-        pub seed: u64,
+        pub seed: u32,
         pub dims: Vec<usize>,
         pub mines: usize,
         pub turn_num: usize,
